requirements.txt
discord.py==2.3.2
tortoise-orm==0.20.0
